build fix

+++ code/qcommon/cm_polylib.cpp
@@ -20,6 +20,7 @@ This file is part of Jedi Academy.
 
 
 #include "cm_local.h"
+#include <stdint.h>
 
 
 // counters are only bumped when running single threaded,
@@ -267,7 +268,7 @@ winding_t	*CopyWinding (winding_t *w)
 	winding_t	*c;
 
 	c = AllocWinding (w->numpoints);
-	size = (int)((winding_t *)0)->p[w->numpoints];
+	size = (intptr_t)((winding_t *)0)->p[w->numpoints];
 	memcpy (c, w, size);
 	return c;
 }
